Fix an bug for wrong cursorposition
The font parameter is used to calculate the visible width of texts. After creation of a textfield it is never changed. If the font is changed this may lead to wrong calculations of text width and things that are based on this calculation (e.g. cursor position, selected areas etc.)