Whole Listbox cells clickable
As discussed in 
https://hub.jmonkeyengine.org/t/lemur-listbox-area-of-clickbehaviour-and-layout-question/40916/6
this bug fix will add an invisible background to the Button in listboxes. That will allow the user to click the whole line in a listbox to achieve a selection. Otherwise he will be only able to select a line if he clicks the text area in listboxes.