Repaired an Bug in Styling
During creation of lemur elements,
while calling 
styles.applyStyles(this, ElementId , style);
the function 
public void initializeStyles(Class c, String style) {
is called. That function is first of all initialisating the styles by checking for an function annotated with @Styledefault in the lemur element class  
"@StyleDefaults(ELEMENT_ID) 
public static void initializeDefaultStyles( Styles styles, Attributes attrs ) {"

Before the changes Attributes in variable attrs always belonged to style „null“ so setting any style in that function like would only apply to the default style or „root“ style in the end

   attrs.set("scrollOnHover",true,false);
To set and apply any options in an element that is not of the current standard style a workaround and passing around variables would have been of need. 
e.g. styles.getSelector(ElemenmtID, Style .set("scrollOnHover", true, false);
The applied changes are minimal and I feel the Style.class as well as the Lemur concept is thought to work that way → thus it was just a BUG